+++ NEWS.md
@@ -1,7 +1,10 @@
+# naryn 2.6.10 
+
+* `emr_track.rm` can now accept multiple tracks as input. 
+
 # naryn 2.6.9 
 
 * Removed global variables from the package - EMR_ROOTS, EMR_GROOT, EMR_UROOT, EMR_FILTERS and EMR_VTRACKS are now a part of a dedicated environment called `.naryn`. You can access them using `.naryn::EMR_ROOTS` etc. 
-
 
 # naryn 2.6.8